tests/requirements.txt
coverage
djangocms_helper
factory_boy
